Replacing in memory hash maps with redis persistence layer

+++ src/main/java/benjamin/groehbiel/ch/api/ShortenerStats.java
@@ -2,13 +2,13 @@ package benjamin.groehbiel.ch.api;
 
 public class ShortenerStats {
 
-    private Integer counter;
+    private Long counter;
 
-    protected ShortenerStats(Integer counter) {
+    protected ShortenerStats(Long counter) {
         this.counter = counter;
     }
 
-    public Integer getCounter() {
+    public Long getCounter() {
         return counter;
     }
 }
